% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/c_mean.R
\name{c_mean}
\alias{c_mean}
\title{Calculate the mean of a numeric vector}
\usage{
c_mean(x)
}
\arguments{
\item{X}{A numeric vector}
}
\value{
The mean of the input vector
}
\description{
This function calculates the mean of a numeric vector
}
\examples{
c_mean(c(1, 2, 3, 4, 5))
}
